changes thanks to code review
